C11309: Extra breaklines added to avoid localization issues
Hello, @ScottAzure,

Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version. Blank spaces added need to be erased, and breaklines need to be added. 

Please, help to check my comment tag added into the article and reply with explanation if source text should be fixed, or it is by design. If fix is needed, please, let me know either if you would like me to fix it in this PR, or if you will fix it in another PR. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.