LOCBLOCKER: L12619: Remove line breaks in content
Hello, @srinathvasireddy, cc @aarzh-AaronZhang

Localization team has reported source content issue that causes localizable text to be blocked from localization. 
We found that content item with line breaks within double quotes prevents the loc automation tool from correctly parsing the localizable strings and causes build break after handback. In order to resolve this issue, we need to remove the line breaks within the content item as suggested in this PR.
The content is metadata.

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.

Cc: @carolinacmoravia @SunnyDeng