C13071: Missing hyperlink due to incorrect asterisk in link structure
Hello, @v-bllydi,

Localization team has reported source content issue that causes issue for localization.

Please, help to check my comment tag added into the article and reply with explanation if source text should be fixed, or it is by design. If fix is needed, please, let me know either if you would like me to fix it in this PR, or if you will fix it in another PR. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.