Bulk edit to get all articles to conform to a standard
I believe I got most of this correct. This is being done as part of an experiment to ensure that the articles are meeting a basic standard and are present for support topics. if you see major formatting issues, most formatting issues existed prior. Example: ## being used twice in an article.

Will be updating content on the next wave. Feel free to point out missing <br>, formatting, white space, etc.