M31119: Extra word at line 30
@krbain
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version. 
Description: Source bug? "admins" does not seem to be necessary. I translated this sentence as such.
Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.
Many thanks in advance.