M32037: Sentence requires clarification
Hello, @jlian,  
Translator has reported possible source content issue. 
Description: Not sure the exact meaning of this sentence. Seems to be grammatically incorrect.
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.