Created common solutions for all of the new support topics
The new proposed support tree is currently being reviewed to be published. These are the common soluitons for all the new support topics. Right now they are basic with just links for the most part so that we at least present something to the customer. The next step we will do is to improve these common solutions.