M38265: Fixing typo 700/7000
Hi @simonschlegel , A previous [PR](https://github.com/Azure/SelfHelpContent/pull/4793) with this change was previously closed. 
could you please confirm if this fix wasn't right? it is fine 700 instead of 7000 even if the filename (and metadata info) is problemscopingques-7000-8000-migration.md?
Please, let us know so we can notify it to the translator