Update hosttoguest-portscanner.md
Remove brackets []. These are no needed for string replacement. From dev Emily Horton:

I think there's a slight issue with the article string replacement. it looks like its written in the article as "<!--$VMArray-->[VMArray]<!--/$VMArray-->", but that signals ADS it's looking for a variable named [VMArray], which can't be done in C# (brackets aren't allowed in variable names). The insight is working correctly otherwise, it's just not able to insert those variables due to those brackets. My variables are called VMArray and PortsArray